+++ burger-queen/src/pages/Login/Login.jsx
@@ -1,20 +1,20 @@
 import React, { useState } from "react";
-import logo from "../img/logo2.png";
-import "../css/Login.css";
+import logo from "../../img/logo2.png";
+import style from "./Login.module.css";
 import axios from "axios";
 import { useNavigate } from "react-router-dom";
-import {Modal} from "../components/Modal/Modal";
+import { Modal } from "../../components/Modal/Modal";
 
 function Login() {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
-  const [modalVisibility, setModalVisibility] = useState(false)
+  const [modalVisibility, setModalVisibility] = useState(false);
   const navigate = useNavigate();
 
   const handleSubmit = (e) => {
     e.preventDefault();
     if (email === "") {
-      setModalVisibility(true)
+      setModalVisibility(true);
     }
 
     const API_URL = "http://localhost:3001/auth";
@@ -31,11 +31,11 @@ function Login() {
   };
 
   return (
-    <section className="login">
-      <div className="background-logo">
-        <img src={logo} alt="burger-queen-logo" className="logo" />
+    <section className={style.login}>
+      <div className={style.backgroundLogo}>
+        <img src={logo} alt="burger-queen-logo" className={style.logo} />
       </div>
-      <div className="background-form">
+      <div className={style.backgroundForm}>
         <h1> BURGER QUEEN </h1>
         <form onSubmit={handleSubmit}>
           <p> Ingresa tus credenciales </p>
@@ -53,7 +53,6 @@ function Login() {
           />
           <input type="submit" value="Entrar" />
         </form>
-        
       </div>
       <Modal state={modalVisibility} onChangeState={setModalVisibility}>
         <h2>Error</h2>
