+++ burger-queen/src/pages/Login/Login.module.css
@@ -7,8 +7,8 @@
     background: #FEDDB0;
 }
 
-.background-logo,
-.background-form {
+.backgroundLogo,
+.backgroundForm {
     display: flex;
     flex-direction: column;
     justify-content: center;
@@ -16,7 +16,7 @@
     padding: 4rem;
 }
 
-.background-logo {
+.backgroundLogo {
     background: #fff;
 }
