+++ burger-queen/src/pages/Waiter/Waiter.jsx
@@ -1,12 +1,12 @@
 /* eslint-disable jsx-a11y/alt-text */
 import React, { useState, useEffect } from "react";
-import Header from "../components/Header/Header.jsx";
-import OrderSheet from "../components/Sheets/OrderSheet.jsx";
-import style from "../css/Waiter.module.css";
-import Card from "../components/Card.jsx";
-import peticionHTTP from "../functions/getProducts";
-import mesero from "../img/mesero.png";
-import NavWaiter from "../components/Nav/NavWaiter.jsx";
+import Header from "../../components/Header/Header.jsx";
+import OrderSheet from "../../components/Sheets/OrderSheet.jsx";
+import style from "./Waiter.module.css";
+import Card from "../../components/Card.jsx";
+import peticionHTTP from "../../functions/getProducts";
+import mesero from "../../img/mesero.png";
+import NavWaiter from "../../components/Nav/NavWaiter.jsx";
 
 export default function Waiter() {
   const [products, setProducts] = useState([]);
